Move definitions around
